pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["Faisal Abuzaid <faisalabuzaid@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.7.1"
whitenoise = "^5.2.0"
httpie = "^2.4.0"
django-cors-headers = "^3.7.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
requests = "^2.25.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
